+++ marklogic-client-api/src/main/java/com/marklogic/client/impl/FnExprImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 MarkLogic Corporation
+ * Copyright (c) 2022 MarkLogic Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.marklogic.client.impl;
 
 import com.marklogic.client.type.XsAnyAtomicTypeSeqVal;
@@ -276,6 +277,18 @@ class FnExprImpl implements FnExpr {
   @Override
   public ServerExpression currentTime() {
     return new XsExprImpl.TimeCallImpl("fn", "current-time", new Object[]{  });
+  }
+
+  
+  @Override
+  public ServerExpression dateTime(ServerExpression arg1, String arg2) {
+    return dateTime(arg1, (arg2 == null) ? (ServerExpression) null : xs.time(arg2));
+  }
+
+  
+  @Override
+  public ServerExpression dateTime(ServerExpression arg1, ServerExpression arg2) {
+    return new XsExprImpl.DateTimeCallImpl("fn", "dateTime", new Object[]{ arg1, arg2 });
   }