+++ marklogic-client-api/src/main/java/com/marklogic/client/impl/RowManagerImpl.java
@@ -256,7 +256,7 @@ public class RowManagerImpl
   }
 
   @Override
-  public <T extends XMLReadHandle> T generateView(Plan plan, String schema, String view, T resultsHandle) {
+  public <T extends XMLReadHandle> T generateView(PlanBuilder.PreparePlan plan, String schema, String view, T resultsHandle) {
     if (resultsHandle == null) {
       throw new IllegalArgumentException("Must specify a handle to generate a view for the plan");
     } else if (schema == null || schema.length() == 0) {
@@ -276,9 +276,35 @@ public class RowManagerImpl
     return services.postResource(requestLogger, "rows", null, params, astHandle, resultsHandle);
   }
   @Override
-  public <T> T generateViewAs(Plan plan, String schema, String view, Class<T> as) {
+  public <T> T generateViewAs(PlanBuilder.PreparePlan plan, String schema, String view, Class<T> as) {
     ContentHandle<T> handle = handleFor(as);
     if (generateView(plan, schema, view, (XMLReadHandle) handle) == null) {
+      return null;
+    }
+
+    return handle.get();
+  }
+
+  @Override
+  public <T extends JSONReadHandle> T columnInfo(PlanBuilder.PreparePlan plan, T resultsHandle) {
+    if (resultsHandle == null) {
+      throw new IllegalArgumentException("Must specify a handle to generate a view for the plan");
+    }
+
+    PlanBuilderBaseImpl.RequestPlan requestPlan = checkPlan(plan);
+    AbstractWriteHandle astHandle = requestPlan.getHandle();
+
+    RequestParameters params = new RequestParameters();
+    params.add("output",     "columnInfo");
+
+    return services.postResource(requestLogger, "rows", null, params, astHandle, resultsHandle);
+  }
+  @Override
+  public <T> T columnInfoAs(PlanBuilder.PreparePlan plan, Class<T> as) {
+    ContentHandle<T> handle = handleFor(as);
+    if (!(handle instanceof JSONReadHandle))
+      throw new IllegalArgumentException("The handle is not an instance of JSONReadHandle.");
+    if (columnInfo(plan, (JSONReadHandle) handle) == null) {
       return null;
     }